+++ app/src/main/java/woowacourse/shopping/data/ProductRemoteMockRepositoryImpl.kt
@@ -5,12 +5,12 @@ import com.example.domain.model.Product
 import com.example.domain.repository.ProductRepository
 
 class ProductRemoteMockRepositoryImpl(
-    private val webServer: ProductMockWebServer,
+    private val service: ProductMockService,
     override val productCache: ProductCache
 ) : ProductRepository {
     override fun getFirstProducts(onSuccess: (List<Product>) -> Unit) {
         if (productCache.productList.isEmpty()) {
-            webServer.request(
+            service.request(
                 page = 1,
                 onSuccess = {
                     productCache.addProducts(it)
@@ -25,7 +25,7 @@ class ProductRemoteMockRepositoryImpl(
 
     override fun getNextProducts(onSuccess: (List<Product>) -> Unit) {
         val currentPage = (productCache.productList.size - 1) / LOAD_SIZE + 1
-        webServer.request(
+        service.request(
             currentPage + 1,
             onSuccess = {
                 productCache.addProducts(it)
